requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pandas>=1.5.0
numpy>=1.23.0
torch>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0

